+++ cmd/zetae2e/local/local.go
@@ -501,6 +501,11 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 	// Verify that the balance of restricted address is zero
 	deployerRunner.EnsureZeroBalanceOnRestrictedAddressZEVM()
 
+	if !deployerRunner.IsRunningUpgrade() {
+		// Verify that there are no stale ballots left over after tests complete
+		deployerRunner.EnsureNoStaleBallots()
+	}
+
 	// print and validate report
 	networkReport, err := deployerRunner.GenerateNetworkReport()
 	if err != nil {